pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image-platform"
version = "0.1.0"
description = "Image platform project"
requires-python = ">=3.11"
dependencies = [
    "openai",
    "gradio",
    "pillow",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "mypy",
    "ruff",
]

[tool.hatch.envs.default]
dependencies = [
    "pip",
]

[tool.hatch.build.targets.wheel]
packages = ["image_platform"]